+++ CostOfRevenue/Models/IIngredient.cs
@@ -14,6 +14,6 @@ namespace CostOfRevenue.Models
         string Id { get; }
         string Name { get; }
         Enums.Unit Unit { get; }
-        float UnitPrice { get; }
+        decimal UnitPrice { get; }
     }
 }
